[UPDATE] Security Problem and Solution (Part 2) - Only owner can mark a task as complete and/or delete it

+++ taskmate/todolist_app/views.py
@@ -29,7 +29,10 @@ def todolist(request):
 @login_required    
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
+    if task.manage == request.user:
+        task.delete()
+    else:
+        messages.error(request,("Access Restricted, You Are Not Allowed!"))
 
     return redirect('todolist')
 
@@ -46,12 +49,16 @@ def edit_task(request, task_id):
     else:
         task_obj = TaskList.objects.get(pk=task_id)
         return render(request, 'edit.html', {'task_obj': task_obj})
+
 @login_required    
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True
-    task.save()
-
+    if task.manage == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request,("Access Restricted, You Are Not Allowed!"))
+    
     return redirect('todolist')
 
 @login_required
